requirements.txt
allure-behave==2.13.1
allure-python-commons==2.13.1
behave==1.2.6
openpyxl==3.1.2
python-dotenv==1.0.0
pywin32==306
requests==2.28.2
selenium==4.8.3
virtualenv==20.21.0
webdriver-manager==3.8.6
